[Android] 특허 통신

See merge request s12-bigdata-recom-sub1/S12P21D208!130

+++ mr_patent_android/app/src/main/java/com/ssafy/mr_patent_android/data/model/response/PatentListResponse.kt
@@ -9,7 +9,7 @@ data class PatentListResponse (
         @SerializedName("patent_draft_id") val patentDraftId: Int,
         @SerializedName("patent_draft_title") val patentDraftTitle: String,
         @SerializedName("fitness_is_corrected") val fitnessIsCorrected: Boolean,
-        @SerializedName("detailed_comparison_total_score") val detailedComparisonTotalScore: Int,
+        @SerializedName("detailed_comparison_total_score") val detailedComparisonTotalScore: Double,
         @SerializedName("created_at") val createdAt: String
     )
 }
